refactor: update CLI implementation and replace CommonJS with ES module syntax
