+++ bin/doctor
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 
-import { cli } from '../dist/cli.cjs';
+import { cli } from '../dist/cli.js';
 await cli(process.argv);
